ai-engine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
openai>=1.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
tree-sitter==0.20.1
tree-sitter-languages==1.8.0
networkx>=3.2.1
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.25.1
numpy<2
langchain
langchain-community
langchain-groq